+++ src/main/components/ShowSelectionAsQrCode.tsx
